ai_fitness_tracker/backend/requirements.txt
Django==4.2.9
djangorestframework==3.14.0
djangorestframework-simplejwt==5.3.1
django-cors-headers==4.3.1
django-filter==23.5
anthropic==0.18.1
Pillow==10.2.0
python-decouple==3.8
gunicorn==21.2.0
